refactor(npc): split lib.rs + manager.rs into focused submodules (#1200)
Behavior-preserving decomposition of the two largest parish-npc files:

- lib.rs (2,584 LOC) -> 79-LOC facade + 9 submodules: npc, response,
  language, tier1_prompt, names, repetition, context, tests, test_helpers.
  Resolves TD-011 (tier1 inline format! prompt now in tier1_prompt) and
  TD-027 (crate-root concern split).
- manager.rs (1,531 LOC) -> manager/ dir: mod (struct + core impl + free
  helpers), intro, lookup, tiers, tests. Resolves TD-030.

Every public path is preserved via lib.rs re-exports; the single
NpcManager impl is split across multiple impl blocks. 526 tests pass
unchanged; architecture-fitness green.
