+++ parish/crates/parish-npc/src/npc.rs
@@ -1,0 +1,153 @@
+//! NPC model: identity, schedule lookups, and the test fixture.
+
+use super::*;
+
+/// A non-player character in the game world.
+///
+/// Contains identity, personality, location, schedule, relationships,
+/// and short-term memory. Cognition fidelity is determined by the
+/// NpcManager based on distance from the player.
+#[derive(Debug, Clone)]
+pub struct Npc {
+    /// Unique identifier.
+    pub id: NpcId,
+    /// Full name.
+    pub name: String,
+    /// Brief anonymous description shown before the player is introduced.
+    ///
+    /// E.g., "a priest", "a middle-aged woman", "an older man".
+    pub brief_description: String,
+    /// Age in years.
+    pub age: u8,
+    /// Occupation or role in the parish.
+    pub occupation: String,
+    /// Personality description used in system prompts.
+    pub personality: String,
+    /// Pronouns used when narrating this NPC (e.g. `he/him`, `she/her`,
+    /// `they/them`). Defaults to `they/them` for NPCs that don't state them
+    /// (#1026).
+    pub pronouns: String,
+    /// Multidimensional intelligence profile shaping dialogue generation.
+    pub intelligence: Intelligence,
+    /// Current location.
+    pub location: LocationId,
+    /// Current emotional state.
+    pub mood: String,
+    /// Home location (where the NPC sleeps).
+    pub home: Option<LocationId>,
+    /// Workplace location (where the NPC works).
+    pub workplace: Option<LocationId>,
+    /// Season- and day-aware schedule defining where the NPC goes.
+    pub schedule: Option<SeasonalSchedule>,
+    /// Relationships to other NPCs, keyed by their id.
+    pub relationships: HashMap<NpcId, Relationship>,
+    /// Ring buffer of recent memories.
+    pub memory: ShortTermMemory,
+    /// Persistent long-term memory with keyword-based retrieval.
+    pub long_term_memory: LongTermMemory,
+    /// Things this NPC knows (local gossip, history, etc.).
+    pub knowledge: Vec<String>,
+    /// Whether the NPC is present at their location or in transit.
+    pub state: NpcState,
+    /// Compact summary from the last tier deflation, if any.
+    ///
+    /// Set when the NPC drops to a lower cognitive tier; cleared when
+    /// they are inflated back to a higher tier.
+    pub deflated_summary: Option<NpcSummary>,
+    /// Log of recent player reactions (emoji) toward this NPC.
+    pub reaction_log: ReactionLog,
+    /// Last activity summary from Tier 3 batch simulation.
+    ///
+    /// Used in deflated context and Tier 3 prompt construction.
+    /// Updated each time a Tier 3 tick processes this NPC.
+    pub last_activity: Option<String>,
+    /// Whether the NPC is currently ill. Set by Tier 4 rules engine.
+    pub is_ill: bool,
+    /// Game-time at which this NPC is fated to die, if set.
+    ///
+    /// Populated by the Tier 4 rules engine when it rolls a `Death` event —
+    /// rather than removing the NPC immediately, the doom is scheduled a few
+    /// game-hours ahead so that [`crate::banshee`] can herald it with a
+    /// keening cry on the night beforehand. Cleared on removal.
+    pub doom: Option<chrono::DateTime<chrono::Utc>>,
+    /// `true` once the banshee's cry has been emitted for the current [`Self::doom`].
+    ///
+    /// Prevents the same wail from being produced on every tick while the
+    /// doom window is open. Reset to `false` whenever [`Self::doom`] changes.
+    pub banshee_heralded: bool,
+}
+
+impl Npc {
+    /// Creates a test NPC for Phase 1 development.
+    ///
+    /// Padraig O'Brien is a 58-year-old publican at The Crossroads,
+    /// known for his storytelling and dry wit.
+    pub fn new_test_npc() -> Self {
+        Self {
+            id: NpcId(1),
+            name: "Padraig O'Brien".to_string(),
+            brief_description: "an older man behind the bar".to_string(),
+            age: 58,
+            occupation: "Publican".to_string(),
+            personality: "A gruff but warm-hearted publican who has run the crossroads \
+                pub for thirty years. Known for his dry wit, encyclopedic knowledge of \
+                local history, and tendency to offer unsolicited advice. He speaks with \
+                a thick Roscommon accent and peppers his speech with Irish phrases."
+                .to_string(),
+            pronouns: "he/him".to_string(),
+            intelligence: Intelligence::new(3, 3, 4, 4, 5, 4),
+            location: LocationId(1),
+            mood: "content".to_string(),
+            home: None,
+            workplace: None,
+            schedule: None,
+            relationships: HashMap::new(),
+            memory: ShortTermMemory::new(),
+            long_term_memory: LongTermMemory::new(),
+            knowledge: Vec::new(),
+            state: NpcState::default(),
+            deflated_summary: None,
+            reaction_log: ReactionLog::default(),
+            last_activity: None,
+            is_ill: false,
+            doom: None,
+            banshee_heralded: false,
+        }
+    }
+
+    /// Returns the name to display to the player.
+    ///
+    /// Before the NPC is introduced, returns the brief anonymous description
+    /// (e.g., "a priest"). After introduction, returns the full name.
+    pub fn display_name(&self, introduced: bool) -> &str {
+        if introduced {
+            &self.name
+        } else {
+            &self.brief_description
+        }
+    }
+
+    /// Returns the NPC's desired location based on their schedule and the current context.
+    ///
+    /// Returns `None` if the NPC has no schedule or no entry covers the hour.
+    pub fn desired_location(
+        &self,
+        hour: u8,
+        season: Season,
+        day_type: DayType,
+    ) -> Option<LocationId> {
+        self.schedule.as_ref()?.location_at(hour, season, day_type)
+    }
+
+    /// Returns the active schedule entry for the current context.
+    ///
+    /// Returns `None` if the NPC has no schedule or no entry covers the hour.
+    pub fn schedule_entry(
+        &self,
+        hour: u8,
+        season: Season,
+        day_type: DayType,
+    ) -> Option<&types::ScheduleEntry> {
+        self.schedule.as_ref()?.entry_at(hour, season, day_type)
+    }
+}
